feature: restructure folders & files
- Removed src/ useless folder from lib/src/
